requirements.txt
anthropic>=0.40.0
requests>=2.32.0
python-dotenv>=1.0.0
gspread>=6.0
google-auth>=2.0
